TestRunState: re-working how we annotate crashing / not run tests
All test runs are now porcessed through TestRunState, even if we have no
reason to believe they crashed with an uncaught signal.  This means that
if the test output gets interrupted for whatever reason, we'll generate
events for the missing portion.

The output annotations are mostly the same as they were before with 1
difference.  If the tests crash before a test runs, or in between tests,
we no longer shove the crash reports in the next test.  Instead, we
always create a placeholder test to hold info on the crash.