we no longer care about otest's exit status or if it had uncaught signals.
We used to only say a test bundle passed if the exit status of
otest/xctest was zero.  And, we'd only do the post-processing of test
events (via OCUnitCrashFilter/TestRunState) if we noticed an uncaught
signal from the child process.

This had a couple problems:  1) we never had a 100% reliable way to
capture the exit status of iOS application tests since the simulator
launched the process (not us);  2) if test output stopped mid-stream but
the test runner didn't crash because of an uncaught signal, we would not
emit all the fake events to pretend the test run finished.

We now run everything through TestRunState, all the time.  Only if
TestRunState indicates everything passed do we say a test bundle passed.