make otest-query work with OS X application tests.
The OS X app test query process now mirrors the iOS one.  We load the
TEST_HOST, inject in otest-query-lib-osx, then have that dlopen() the
test bundle, get the tests, then exit.

We have to do this because the test bundle may have dependencies on
symbols inside the TEST_HOST, and so it can't always be loaded by
itself.