to set the simulator device, use setSimulatedDeviceInfoName:
This removes the hack where we would set the 'SimulateDevice'
user-default as a way of selecting the simulator device.

NOTE: This moves us to a broken state in the interim.  For now, the only
way to run application tests is to supply -destination.  It will fail
without that.

Later diffs will fix this problem.