pull DestinationInfo thru to the end; it knows cpuType + deviceName
We used to have a bunch of extra plumbing to carry the deviceName and
cpuType values all way into the OCUnitTestRunner and
TestableExecutionInfo.  Now we just pass along DestinationInfo.

NOTE: This temporarily breaks running 64-bit tests, as
`-[DestinationInfo cpuTypeOrImpliedCpuType]` will always return i386.
We'll fix this in a later diff.